@@ -1,9 +1,1 @@
-function showLoadingOverlay() {
-    const overlay = document.getElementById('loadingOverlay');
-    overlay.classList.remove('d-none');
-}
-
-function hideLoadingOverlay() {
-    const overlay = document.getElementById('loadingOverlay');
-    overlay.classList.add('d-none');
-}
+import './bootstrap';
